Conflicts:
	watchDir/new_server.py

+++ alexaNewCode/serverConnections.py
@@ -1,5 +1,6 @@
 
 from serverClasses import *
+import sys
 
 from multiprocessing import Process
 
@@ -24,6 +25,7 @@ class TransferCancelled(Exception):
     """ Exception for a user cancelling a transfer """
     pass
 
+
 class ServerReceiverProtocol(basic.LineReceiver):
     """ File Receiver """
 
@@ -32,21 +34,34 @@ class ServerReceiverProtocol(basic.LineReceiver):
         self.outfile = None
         self.remain = 0
         self.crc = 0
+        self.isFile = True
 
     def lineReceived(self, line):
         """ """
         print ' ~ lineReceived:\n\t', line
-        self.instruction = json.loads(line)
-        self.instruction.update(dict(client=self.transport.getPeer().host))
-        self.size = self.instruction['file_size']
-        self.original_fname = self.instruction.get('original_file_path',
-                                                   'not given by client')
+
+        self.isFile = True
 
         # get username via ip address
         clientUsername = self.factory.retrieveUser(self.transport.getPeer().host)
 
         # Create the upload directory if not already present
         uploaddir = os.path.join(self.factory.dir_path, clientUsername)
+
+        if line[0:6] == 'delete':
+            originalPath = line[6:]
+            pathToDelete = os.path.join(uploaddir, os.path.basename(originalPath))
+            os.remove(pathToDelete)
+            self.isFile = False
+            self.transport.loseConnection()
+            return
+
+        self.instruction = json.loads(line)
+        self.instruction.update(dict(client=self.transport.getPeer().host))
+        self.size = self.instruction['file_size']
+        self.original_fname = self.instruction.get('original_file_path',
+                                                   'not given by client')
+
 
         print " * Using upload dir:",uploaddir
         if not os.path.isdir(uploaddir):
@@ -83,32 +98,35 @@ class ServerReceiverProtocol(basic.LineReceiver):
         print ' * ',self.transport.getPeer()
 
     def connectionLost(self, reason):
-        """ """
-        basic.LineReceiver.connectionLost(self, reason)
-        print ' - connectionLost'
-        if self.outfile:
-            self.outfile.close()
-            # Problem uploading - tmpfile will be discarded
-        if self.remain != 0:
-            print str(self.remain) + ')!=0'
-            remove_base = '--> removing tmpfile@'
-            if self.remain<0:
-                reason = ' .. file moved too much'
-            if self.remain>0:
-                reason = ' .. file moved too little'
-            print remove_base + self.outfilename + reason
-            os.remove(self.outfilename)
 
-        # Success uploading - tmpfile will be saved to disk.
-        else:
+        if self.isFile == False:
+            print 'connection lost'
+        if self.isFile == True:
+            basic.LineReceiver.connectionLost(self, reason)
+            print ' - connectionLost'
+            if self.outfile:
+                self.outfile.close()
+                # Problem uploading - tmpfile will be discarded
+            if self.remain != 0:
+                print str(self.remain) + ')!=0'
+                remove_base = '--> removing tmpfile@'
+                if self.remain<0:
+                    reason = ' .. file moved too much'
+                if self.remain>0:
+                    reason = ' .. file moved too little'
+                print remove_base + self.outfilename + reason
+                os.remove(self.outfilename)
 
-            user_address = self.transport.getPeer().host
+            # Success uploading - tmpfile will be saved to disk.
+            else:
 
-            self.factory.sendToMachines(user_address, self.outfilename)
+                #user_address = self.transport.getPeer().host
 
-            print '\n--> finished saving upload@' + self.outfilename
+                #self.factory.sendToMachines(user_address, self.outfilename)
 
-            #self.factory.testSendMachines(user_address)
+                print '\n--> finished saving upload@' + self.outfilename
+
+                #self.factory.testSendMachines(user_address)
 
 def fileinfo(fname):
     """ when "file" tool is available, return it's output on "fname" """
@@ -213,7 +231,7 @@ class FileIOClientFactory(ClientFactory):
         return p
 
 
-def transmitOne(path, address='localhost', port=1235,):
+def transmitOne(path, address=str(sys.argv[1]), port=1235,):
     """ helper for file transmission """
     controller = type('test',(object,),{'cancel':False, 'total_sent':0,'completed':Deferred()})
     f = FileIOClientFactory(path, controller)
@@ -226,7 +244,7 @@ class FileIOServerFactory(ServerFactory):
     """ file receiver factory """
     protocol = ServerReceiverProtocol
 
-    def __init__(self, filePath, address='localhost', send_port=1235, listen_port=1234):
+    def __init__(self, filePath, address=str(sys.argv[1]), send_port=1235, listen_port=1234):
         """ """
         # server filepath
         self.dir_path = filePath
@@ -238,7 +256,7 @@ class FileIOServerFactory(ServerFactory):
 
         self.adminUser = ["admin"]
         self.usersToPW = {"admin": "pw", "alexa": "14"}
-        adminMachine = BabyLocalMachine("admin", "1", "localhost", "path")
+        adminMachine = BabyLocalMachine("admin", "1", str(sys.argv[1]), "path")
         self.usersToLM = {"admin": [adminMachine]}
 
 
@@ -270,11 +288,15 @@ class FileIOServerFactory(ServerFactory):
         reactor.connectTCP(address, port, f)
         return controller.completed
 
+
 if __name__ == "__main__":
 
-    HOST, PORT = "localhost", 9999
+    HOST, PORT = str(sys.argv[1]), 9999
 
-    twisted_server = FileIOServerFactory('/Users/alowman/TestServer')
+    print HOST
+    print PORT
+
+    twisted_server = FileIOServerFactory(str(sys.argv[2]))
 
     # Create the server, binding to localhost on port 9999
     socket_server = ThreadedTCPServer((HOST, PORT), MyTCPHandler, twisted_server)
